refactor: remove tests in cli module files
- we remove some tests that got duplicated during the refactor of the CLI
- the remaining tests will be moved into the integration tests added in #353
